+++ tests/model/pyarrow/test_pyarrow_aggr_dataset.py
@@ -3,6 +3,7 @@ import pytest
 
 from toshi_hazard_store.model.hazard_models_pydantic import HazardAggregateCurve
 from toshi_hazard_store.model.pyarrow import pyarrow_aggr_dataset, pyarrow_dataset
+from toshi_hazard_store.model.pyarrow.dataset_schema import get_hazard_aggregate_schema
 
 
 @pytest.fixture
@@ -15,7 +16,7 @@ def random_hazard_curves():
                 nloc_001="-100.100~45.045",
                 nloc_0="-100.0~45.0",
                 imt="PGA",
-                vs30=1000,
+                vs30="1000",
                 aggr="mean",
                 values=[(x / 1000) for x in range(44)],
             )
@@ -34,11 +35,14 @@ def test_serialise_aggregate_hazard_curves(tmp_path, random_hazard_curves):
     )
 
     # read and check the dataset
-    dataset = ds.dataset(output_folder, format='parquet', partitioning='hive')
+    schema = get_hazard_aggregate_schema()
+    dataset = ds.dataset(output_folder, format='parquet', partitioning='hive', schema=schema)
     table = dataset.to_table()
     dfout = table.to_pandas()
 
+    print("out df >>>")
     print(dfout.shape)
     print(dfout.tail())
     print(dfout.info())
+    print("out df <<<")
     assert dfout.shape == (10, 8)