+++ toshi_hazard_store/model/pyarrow/dataset_schema.py
@@ -28,16 +28,17 @@ def get_hazard_realisation_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEF
     vtype = pyarrow.float64() if use_64_bit_values else pyarrow.float32()
     values_type = pyarrow.list_(vtype)
     vs30_type = pyarrow.int32()
-    dict_type = pyarrow.dictionary(pyarrow.int32(), pyarrow.string(), True)
+    dict_type = pyarrow.dictionary(pyarrow.int8(), pyarrow.string(), False)
+    str_type = pyarrow.string()
 
     return pyarrow.schema(
         [
-            ("compatible_calc_id", dict_type),  # for hazard-calc equivalence, for PSHA engines interoperability
+            ("compatible_calc_id", str_type),  # for hazard-calc equivalence, for PSHA engines interoperability
             ("producer_digest", dict_type),  # digest for the producer look up
             ("config_digest", dict_type),  # digest for the job configutation
-            ("calculation_id", dict_type),  # a reference to the original calculation that produced this item
-            ("nloc_001", dict_type),  # the location string to three places e.g. "-38.330~17.550"
-            ("nloc_0", dict_type),  # the location string to zero places e.g.  "-38.0~17.0" (used for partioning)
+            ("calculation_id", str_type),  # a reference to the original calculation that produced this item
+            ("nloc_001", str_type),  # the location string to three places e.g. "-38.330~17.550"
+            ("nloc_0", str_type),  # the location string to zero places e.g.  "-38.0~17.0" (used for partioning)
             ('imt', dict_type),  # the imt label e.g. 'PGA', 'SA(5.0)'
             ('vs30', vs30_type),  # the VS30 integer
             ('rlz', dict_type),  # the rlz id from the the original calculation
@@ -64,20 +65,25 @@ def get_hazard_aggregate_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEFAU
         values: a list of the 44 IMTL values
     """
     # create a schema...
+
     vtype = pyarrow.float64() if use_64_bit_values else pyarrow.float32()
     values_type = pyarrow.list_(vtype)
+
     vs30_type = pyarrow.int32()
-    dict_type = pyarrow.dictionary(pyarrow.int32(), pyarrow.string(), True)
+
+    # # dict_type = pyarrow.string()
+    # dict_type_int8 = pyarrow.dictionary(pyarrow.int8(), pyarrow.string())
+    # # dict_type_int16 = pyarrow.dictionary(pyarrow.int16(), pyarrow.string())
 
     return pyarrow.schema(
         [
-            ("compatible_calc_id", dict_type),  # for hazard-calc equivalence, for PSHA engines interoperability
-            ("hazard_model_id", dict_type),  # the model that these curves represent.
-            ("nloc_001", dict_type),  # the location string to three places e.g. "-38.330~17.550"
-            ("nloc_0", dict_type),  # the location string to zero places e.g.  "-38.0~17.0" (used for partioning)
-            ('imt', dict_type),  # the imt label e.g. 'PGA', 'SA(5.0)'
+            ("compatible_calc_id", pyarrow.string()),  # for hazard-calc equivalence, for PSHA engines interoperability
+            ("hazard_model_id", pyarrow.string()),  # the model that these curves represent.
+            ("nloc_001", pyarrow.string()),  # the location string to three places e.g. "-38.330~17.550"
+            ("nloc_0", pyarrow.string()),  # the location string to zero places e.g.  "-38.0~17.0" (used for partioning)
+            ('imt', pyarrow.string()),  # the imt label e.g. 'PGA', 'SA(5.0)'
             ('vs30', vs30_type),  # the VS30 integer
-            ('aggr', dict_type),  # the the aggregation type
+            ('aggr', pyarrow.string()),  # the the aggregation type
             ("values", values_type),  # a list of the 44 IMTL values
         ]
     )